Move Addressable into the wasm package
It's only needed for wasm, not anything else.
